add inline toolbar for crosslinks
